% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTimeVaryData.R
\name{addTimeVaryData}
\alias{addTimeVaryData}
\title{Add Time-Varying Data}
\usage{
addTimeVaryData(
  filter_ch = NULL,
  start_date = NULL,
  end_date = NULL,
  strata_beg = "Mon",
  last_strata_min = 3
)
}
\arguments{
\item{filter_ch}{filtered capture history as returned by the `filterDetections()`
function in the `PITcleanr` package, which has then been verified by a user and all
blank or NA `user_keep_obs` entries have been completed.}

\item{start_date}{character vector of date (\code{YYYYMMDD}) when query should start}

\item{end_date}{character vector of date (\code{YYYYMMDD}) when query should end}

\item{strata_beg}{3 letter code for the day of the week each weekly strata should begin on. Default value is \code{'Mon'}.}

\item{last_strata_min}{minimum length (in days) for the final strata. Default value is 3.}
}
\description{
Adds additional data required for the time-varying version of DABOM.
}
\examples{
addTimeVaryData()
}
\author{
Kevin See
}
