% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBranchEscape_GRA.R
\name{calcBranchEscape_GRA}
\alias{calcBranchEscape_GRA}
\title{Estimate Escapement - LGD}
\usage{
calcBranchEscape_GRA(
  dabom_mod = NULL,
  stadem_mod = NULL,
  stadem_param_nm = "X.new.wild",
  bootstrap_samp = 2000
)
}
\arguments{
\item{dabom_mod}{The result of DABOM. An MCMC.list, where the detection parameter names all end with "\code{_p}".}

\item{stadem_mod}{The result of STADEM. An MCMC.list, or \code{jagsUI} object that can be coerced to an MCMC.list.}

\item{stadem_param_nm}{Character vector of length one, with the name of the parameter corresponding to total escapement past Lower Granite Dam.}

\item{bootstrap_samp}{The number of samples to be drawn from the posteriors of the STADEM model and the DABOM model.}
}
\description{
Combines estimates of total escapement from STADEM with transition probabilities from DABOM to generate estimates of escapement above various detection sites. Currently only works for wild fish.
}
\examples{
#calcBranchEscape_GRA()
}
\author{
Kevin See
}
