% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRepGrpEscape_GRA.R
\name{calcRepGrpEscape_GRA}
\alias{calcRepGrpEscape_GRA}
\title{Estimate Escapement for Reporting Groups - GRA}
\usage{
calcRepGrpEscape_GRA(
  dabom_mod = NULL,
  stadem_mod = NULL,
  stadem_param_nm = "X.new.wild",
  bootstrap_samp = 2000,
  parent_child = NULL,
  cred_int_prob = 0.95,
  spp = c("Chinook", "Steelhead")
)
}
\arguments{
\item{dabom_mod}{The result of DABOM. An MCMC.list, where the detection parameter names all end with "\code{_p}".}

\item{stadem_mod}{The result of STADEM. An MCMC.list, or \code{jagsUI} object that can be coerced to an MCMC.list.}

\item{stadem_param_nm}{Character vector of length one, with the name of the parameter corresponding to total escapement past Lower Granite Dam.}

\item{bootstrap_samp}{The number of samples to be drawn from the posteriors of the STADEM model and the DABOM model.}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}

\item{cred_int_prob}{A numeric scalar in the interval (0,1) giving what highest posterior density portion of the posterior the credible interval should cover. The default value is 95\%.}

\item{spp}{Species, either "Chinook" or "Steelhead"}
}
\description{
Combines estimates of total wild escapement at Lower Granite Dam (from STADEM) with transition probabilities from DABOM to generate estimates of escapement to pre-defined reporting groups.
}
\examples{
#calcRepGrpEscape_GRA()
}
\author{
Kevin See
}
