% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRepGrpEscape_LGD.R
\name{calcRepGrpEscape_LGD}
\alias{calcRepGrpEscape_LGD}
\title{Estimate Escapement for Reporting Groups - LGD}
\usage{
calcRepGrpEscape_LGD(dabom_mod = NULL, stadem_mod = NULL,
  stadem_param_nm = "X.new.wild", bootstrap_samp = 2000,
  node_order = NULL, summ_results = T, pt_est_nm = NULL,
  cred_int_prob = 0.95, spp = c("Chinook", "Steelhead"))
}
\arguments{
\item{dabom_mod}{The result of DABOM. An MCMC.list, where the detection parameter names all end with "\code{_p}".}

\item{stadem_mod}{The result of STADEM. An MCMC.list, or \code{jagsUI} object that can be coerced to an MCMC.list.}

\item{stadem_param_nm}{Character vector of length one, with the name of the parameter corresponding to total escapement past Lower Granite Dam.}

\item{bootstrap_samp}{The number of samples to be drawn from the posteriors of the STADEM model and the DABOM model.}

\item{node_order}{output of function \code{createNodeOrder}.}

\item{summ_results}{Should the resulting posteriors be summarised or returned as posterior samples? Default value is \code{TRUE}.}

\item{pt_est_nm}{Determines whether to use the mean, median of mode of posterior samples as the point estimate. Returns all 3 if \code{NULL}, which is default.}

\item{cred_int_prob}{A numeric scalar in the interval (0,1) giving what higest posterior density portion of the posterior the credible interval should cover. The default value is 95\%.}

\item{spp}{Species, either "Chinook" or "Steelhead"}
}
\description{
Combines estimates of total escapement from STADEM with transition probabilities from DABOM to generate estimates of escapement within pre-defined reporting groups. Currently only works for wild fish.
}
\examples{
#calcRepGrpEscape_LGD()
}
\author{
Kevin See
}
