% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinook15_trapdata.R
\docType{data}
\name{chinook15_trapdata}
\alias{chinook15_trapdata}
\title{Lower Grante trap data for spawn Year 2015 Chinook

An example dataset returned from running the \code{LGTrapData} function when
  queried for species = 'Chinook' and spawnyear = '2015'}
\format{A data frame with 22973 rows and 28 variables}
\source{
Idaho Fish and Game's Lower Granite Trapping Database
}
\usage{
chinook15_trapdata
}
\keyword{datasets}

