% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTimeVaryTransProbs.R
\name{compileTimeVaryTransProbs}
\alias{compileTimeVaryTransProbs}
\title{Compile Time-Varying Transition Probabilities}
\usage{
compileTimeVaryTransProbs(dabom_mod = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Extracts the MCMC posteriors of time-varying transition probabilities for a DABOM model. The time-varying parameters should be set up so that they are organized as an array with the first dimension corresponding to the fish origin, the second to the model branch, and the third the model time strata.
}
\examples{
compileTimeVaryTransProbs()
}
\author{
Kevin See
}
