% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTransProbs_GRA.R
\name{compileTransProbs_GRA}
\alias{compileTransProbs_GRA}
\title{Compile Transition Probabilities - GRA}
\usage{
compileTransProbs_GRA(dabom_mod = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model, and multiplies them appropriately. This function is specific to the Lower Granite version of DABOM.
}
\examples{
compileTransProbs_GRA()
}
\author{
Kevin See
}
