% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTransProbs_LGD.R
\name{compileTransProbs_LGD}
\alias{compileTransProbs_LGD}
\title{Compile Transition Probabilities - LGD}
\usage{
compileTransProbs_LGD(dabom_mod = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model, and multiplies them appropriately. This function is specific to the Lower Granite version of DABOM.
}
\examples{
compileTransProbs_LGD()
}
\author{
Kevin See
}
