% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTransProbs_PRA_pre2023.R
\name{compileTransProbs_PRA_pre2023}
\alias{compileTransProbs_PRA_pre2023}
\title{Compile Transition Probabilities - PRA pre 2023}
\usage{
compileTransProbs_PRA_pre2023(dabom_mod = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model, and multiplies them appropriately. This function is specific to the Priest Rapids version of DABOM. Copied from the old version of compileTransProbs_PRA here for all years before 2023 when new sites were not added.
}
\author{
Kevin See
}
