% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTransProbs_PRO.R
\name{compileTransProbs_PRO}
\alias{compileTransProbs_PRO}
\title{Compile Transition Probabilities - PRO}
\usage{
compileTransProbs_PRO(dabom_mod = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}

\item{parent_child}{dataframe created from \code{PITcleanr::createParentChildDf} or similar that contains at least ParentNode, ChildNode and river kilometer columns}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model, and multiplies them appropriately. This function is specific to the Prosser dam version of DABOM.
}
\examples{
compileTransProbs_PRO()
}
\author{
Kevin See
}
