% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTransProbs_TUM.R
\name{compileTransProbs_TUM}
\alias{compileTransProbs_TUM}
\title{Compile Transition Probabilities - TUM}
\usage{
compileTransProbs_TUM(dabom_mod = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model, and multiplies them appropriately. This function is specific to the Tumwater version of DABOM.
}
\examples{
compileTransProbs_TUM()
}
\author{
Kevin See
}
