% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileWeekTransProbs.R
\name{compileWeekTransProbs}
\alias{compileWeekTransProbs}
\title{Compile Weekly Transition Probabilities - LGD}
\usage{
compileWeekTransProbs(dabom_mod = NULL, param_name = "p_pop_main")
}
\arguments{
\item{dabom_mod}{An MCMC.list.}

\item{param_name}{A character vector of length one, containing the name of a parameter that is time-varying by week. Default value is '\code{p_pop_main}'.}
}
\description{
Extracts the MCMC posteriors of time-varying transition probabilities for a DABOM model. The time-varying parameters should be set up so that they are organized with rows corresponding to weeks, and columns corresponding to branches.
}
\examples{
compileWeekTransProbs()
}
\author{
Kevin See
}
