% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{config}
\alias{config}
\title{PIT tag observation site configuration

  An example of a PIT tag site configuration file that maps SiteID, AntennaID and antenna configuration ID to a model node.}
\format{A data frame with 5038 rows and 19 variables}
\source{
\url{http://www.PTAGIS.org} and Rick Orme
}
\usage{
config
}
\keyword{datasets}

