% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDirichletVector.R
\name{createDirichletVector}
\alias{createDirichletVector}
\title{Create Dirichlet vector}
\usage{
createDirichletVector(n_branches = NULL, init_vals = NULL,
  initial_one = T, final_one = F)
}
\arguments{
\item{n_branches}{How many paths could a tag travel from this branching node?}

\item{init_vals}{The initial values, based on observations}

\item{initial_one}{Should the first value of the Dirichlet vector be fixed to 1? Default value is \code{FALSE}.}

\item{final_one}{Should the last value of the Dirichlet vector be fixed to 1? Default value is \code{FALSE}.}
}
\description{
Construct appropriate Dirichlet vector to pass to JAGS, based on which branches had observed tags, and which did not.
}
\examples{
createDirichletVector(3)
}
\author{
Kevin See
}
