% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJAGSinputs_LGR.R
\name{createJAGSinputs_LGR}
\alias{createJAGSinputs_LGR}
\title{Prep DABOM JAGS inputs - LGR}
\usage{
createJAGSinputs_LGR(
  filter_ch = NULL,
  parent_child = NULL,
  configuration = NULL,
  fish_origin = NULL
)
}
\arguments{
\item{filter_ch}{filtered capture history as returned by the `filterDetections()`
function in the `PITcleanr` package, which has then been verified by a user and all
blank or NA `user_keep_obs` entries have been completed.}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`}

\item{fish_origin}{tibble containing columns of `tag_code` and `origin`, where `origin` is either "W" for wild or "H" for hatchery}
}
\description{
Construct all the necessary inputs to the DABOM JAGS model. Returns a named list that can be passed directly to JAGS.
}
\examples{
createJAGSinputs_LGR()
}
\author{
Kevin See
}
