% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJAGSinputs_PRA.R
\name{createJAGSinputs_PRA}
\alias{createJAGSinputs_PRA}
\title{Prep DABOM JAGS inputs - PRA}
\usage{
createJAGSinputs_PRA(dabom_list = NULL)
}
\arguments{
\item{dabom_list}{output of \code{createDABOMcapHist} with parameter split_matrics set to \code{TRUE}.}
}
\description{
Construct all the necessary inputs to the DABOM JAGS model for PRA Returns a named list that can be passed directly to JAGS.
}
\examples{
createJAGSinputs_PRA()
}
\author{
Kevin See
}
