% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJAGSinputs_TUM.R
\name{createJAGSinputs_TUM}
\alias{createJAGSinputs_TUM}
\title{Prep DABOM JAGS inputs - TUM}
\usage{
createJAGSinputs_TUM(dabom_list = NULL, n_branch_list = NULL)
}
\arguments{
\item{dabom_list}{output of \code{createDABOMcapHist} with parameter split_matrics set to \code{TRUE}.}

\item{n_branch_list}{list containing an entry for each branching node, and an integer indicating the number of branches upstream of that node.}
}
\description{
Construct all the necessary inputs to the DABOM JAGS model for TUM. Returns a named list that can be passed directly to JAGS.
}
\examples{
createJAGSinputs_TUM()
}
\author{
Kevin See
}
