% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNodeList.R
\name{createNodeList}
\alias{createNodeList}
\title{Node List}
\usage{
createNodeList(node_order = NULL)
}
\arguments{
\item{node_order}{output of function \code{createNodeOrder}}
}
\description{
Takes the node order dataframe and converts it to a list with an entry for each "group" in the node order dataframe. Each entry in the list contains a character vector of the nodes within that group.
}
\examples{
createNodeList()
}
\author{
Kevin See
}
