% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNodeOrder.R
\name{createNodeOrder}
\alias{createNodeOrder}
\title{Node Order}
\usage{
createNodeOrder(valid_paths = NULL, configuration = NULL,
  parent_child = NULL, site_df = NULL)
}
\arguments{
\item{valid_paths}{a dataframe created by the \code{getValidPaths} function in the \code{PITcleanr} package.}

\item{configuration}{a configuration file built by \code{buildConfig} in the \code{PITcleanr} package.}

\item{parent_child}{is a data frame created by \code{createParentChildDf} in the \code{PITcleanr} package.}
}
\description{
Based on valid paths, returns a dataframe with a row for each node, showing the path to get to that node, and how many nodes a tag must pass to get there.
}
\examples{
createNodeOrder()
}
\author{
Kevin See
}
