% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineRepGrps.R
\name{defineRepGrps}
\alias{defineRepGrps}
\title{Define Report Groups - GRA}
\usage{
defineRepGrps(spp = c("Chinook", "Steelhead"))
}
\arguments{
\item{spp}{Species, either "Chinook" or "Steelhead"}
}
\description{
Define which detection sites fall into which report groups, and what the lowest array is. These reporting groups differ by species.
}
\examples{
defineRepGrps()
}
\author{
Kevin See
}
