% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineRepGrps_GRA.R
\name{defineRepGrps_GRA}
\alias{defineRepGrps_GRA}
\title{Define Report Groups - GRA}
\usage{
defineRepGrps_GRA(spp = c("Chinook", "Steelhead"))
}
\arguments{
\item{spp}{Species, either "Chinook" or "Steelhead"}
}
\description{
Define which detection sites fall into which report groups, and what the lowest array is. These reporting groups differ by species. These are specific to the Lower Granite version of DABOM.
}
\examples{
defineRepGrps_GRA("Chinook")
}
\author{
Kevin See
}
