% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estMode.R
\name{estMode}
\alias{estMode}
\title{Estimate Mode of Posterior Distribution}
\usage{
estMode(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Estimates the mode of the posterior, by calculating the density of posterior samples and choosing x-value with the maximum y-value.
}
\examples{
estMode(rgamma(1000, 5, 1/6))
}
\author{
Kevin See
}
