% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDetectPost.R
\name{extractDetectPost}
\alias{extractDetectPost}
\title{Extract Detection Posteriors}
\usage{
extractDetectPost(dabom_mod = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list containing some parameters that end with "_p"}
}
\description{
Extracts the MCMC posteriors of detection probabilities for a DABOM model
}
\examples{
#extractDetectPost()
}
\author{
Kevin See
}
