% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTransPost.R
\name{extractTransPost}
\alias{extractTransPost}
\title{Extract Transition Posteriors}
\usage{
extractTransPost(dabom_mod = NULL, parent_child = NULL, configuration = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list containing some parameters that begin with "p_pop", "psi" or "phi"}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model
}
\examples{
extractTransPost()
}
\author{
Kevin See
}
