% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTransProbs.R
\name{extractTransProbs}
\alias{extractTransProbs}
\title{Extract Transition Probabilities}
\usage{
extractTransProbs(dabom_mod = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model
}
\examples{
extractTransProbs()
}
\author{
Kevin See
}
