% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishPaths.R
\name{fishPaths}
\alias{fishPaths}
\title{fishPaths: creates a data frame of all observed paths}
\usage{
fishPaths(valid_obs, valid_paths)
}
\arguments{
\item{valid_obs}{}

\item{valid_paths}{}

\item{tagCntLimit}{}
}
\description{
The function builds a data frame of all the observed fish paths in the
observation file based on the parent child table. The function was originally called
FishPath_truncateToLastLeg() and was built by Greg K. Ryan K. altered the function
to work solely in R and doesn't require the SQLite backend.
}
\examples{
fishPaths()

}
\author{
Greg Kliewer
}

