% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNoFishNodes.R
\name{fixNoFishNodes}
\alias{fixNoFishNodes}
\title{Modify basic JAGS model}
\usage{
fixNoFishNodes(init_file = NULL, file_name = NULL, proc_ch = NULL,
  node_order = NULL)
}
\arguments{
\item{init_file}{name (with file path) of basic JAGS model.}

\item{file_name}{name (with file path) to save the model as.}

\item{proc_ch}{capture history as returned by one of the \code{processCapHist} family of functions in \code{PITcleanr} package, which has then been verified by a user and all blank UserProcStatus entries have been completed.}

\item{node_order}{output of function \code{createNodeOrder}}
}
\description{
Takes the basic JAGS model and modifies it based on particular observations. Specifically, it fixes the detection probability of all nodes that had no observations to 0.
}
\examples{
fixNoFishNodes()
}
\author{
Kevin See
}
