% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeInfo_LGR.R
\name{getNodeInfo_LGR}
\alias{getNodeInfo_LGR}
\title{Compile information about each node - LGR}
\usage{
getNodeInfo_LGR(parent_child = NULL, configuration = NULL)
}
\arguments{
\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`}
}
\description{
Based on a parent-child table and a configuration file, this function compiles how many nodes each site has, what the parent site is, and what child number of that parent the site in question is
}
\examples{
writeDABOM()
}
\author{
Kevin See
}
