% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGTrapData.R
\name{LGTrapData}
\alias{LGTrapData}
\title{LGTrapData: retrieve Lower Granite trapping data}
\usage{
LGTrapData(filename, species = c("Chinook", "Steelhead"), spawnyear)
}
\arguments{
\item{filename}{filepath of the trapping database as a character string}
}
\description{
LGTrapData() is a function to query data from the LGTrappingExportJodyW.accdb database
}
\examples{
LGTrapData(filename = './data/Data_Input/LGTrappingExportJodyW.accdb',
species = 'Chinook', spawnyear = '2015')

}
\author{
Ryan N. Kinzer
}

