% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGTrappingDB_03192017.R
\docType{data}
\name{LGTrappingDB_03192017}
\alias{LGTrappingDB_03192017}
\title{LGTrappingDB}
\format{Data from the LGTrappingDB}
\usage{
LGTrappingDB_03192017
}
\description{
This is the Lower Granite Dam Trapping Database exported from the QCI download site on 03-17-19. This
contains all records in the LGTrapping database including both adults and juveniles and from all spawn years.
}
\keyword{Chinook}
\keyword{DABOM}
\keyword{LGR}
\keyword{LGTrappingDB}
\keyword{steelhead}

