% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logOnePath.R
\name{logOnePath}
\alias{logOnePath}
\title{logOnePath: creates a data frame of one path}
\usage{
logOnePath(rootNode, childNode)
}
\arguments{
\item{rootNode}{}

\item{childNode}{}
}
\value{
NULL
#log path from given site to root (mouth of river)
}
\description{
The function builds a data frame of one potential path from the parent child
table. The function was originally called shortestPathToRoot_logOnePath()
and was built by Greg K. Ryan K. altered the function to work solely in R
and to not require the SQLite backend.
}
\examples{
logOnePath()

}
\author{
Greg Kliewer
}

