% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeAssign.R
\name{nodeAssign}
\alias{nodeAssign}
\title{nodeAssign: Assigns PIT-tag observation node site names to PTAGIS tag detection history
records.}
\usage{
nodeAssign(valid_tags, observation, configuration, truncate = FALSE)
}
\arguments{
\item{valid_tags}{is a data frame or comma seperated value (.csv) file containing the
valid tag list which is outputted from the validTagList() function and the LGTrappingDB}

\item{observation}{is the PTAGIS observation file inputted as a data frame or .csv file
containing the complete tag history for each of the tagIDs in valid_tags}

\item{configuration}{is a data frame or .csv file which assigns node names to unique SiteID,
AntennaID, and site configuration ID combinations.}

\item{truncate}{logical, subsets observations to those with valid nodes, observations dates
greater than trapping date at LGD and then to the minimum observation date of each set of
observation events at a node, multiple observation events can occur at one node if the
observations are split by detections at other nodes}
}
\description{
The function assigns PIT-tag observation node site names to each
tag detection history record generated from a PTAGIS 'complete tag history query'.  The
complete tag history query is completed by running the tag list outputted from the
validTagList() function. Observation node site names are assigned from joining a configuration
file with the PTAGIS query results on 'Site_Code' and 'AntennaID' fields. The configuration
file, 'siteDescription.csv', is distributed and maintained within the DABOM package.

An observation node is a single or group of PIT-tag antenna that, at the finest resolution,
act as a unique tag detection location.  Often, a node consists of multiple antennas at a unique
PTAGIS interogation site, which together form a single array.  In addition, multiple nodes
may exist at one PTAGIS interogation site or mark-release-recapture (MRR) site when more than
antenna array are assigned to the same SiteID. A node may also be a single antenna or coil
located in an adult ladder or trap entry or the single (MRR) SiteID for fish handled and
scanned at a weir location.
}
\examples{
nodeAssign()

}
\author{
Ryan Kinzer
}

