% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentchild.R
\docType{data}
\name{parentchild}
\alias{parentchild}
\title{Parent-child relationship of DABOM nodes.

  An example of a parent-child table of model nodes listed in the site configuration file.}
\format{A data frame with 207 rows and 5 variables}
\source{
Rick Orme
}
\usage{
parentchild
}
\keyword{datasets}

