% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepJAGSinputs_LGD.R
\name{prepJAGSinputs_LGD}
\alias{prepJAGSinputs_LGD}
\title{Compile JAGS inputs - Lower Granite Dam}
\usage{
prepJAGSinputs_LGD(proc_ch = NULL, node_order = NULL, time_varying = T,
  spawn_yr = NULL, spp = NULL, start_day = NULL, end_day = NULL,
  strata_beg = "Mon", last_strata_min = 3)
}
\arguments{
\item{proc_ch}{capture history as returned by one of the \code{processCapHist} family of functions in \code{PITcleanr} package, which has then been verified by a user and all blank UserProcStatus entries have been completed.}

\item{node_order}{output of function \code{createNodeOrder}}

\item{time_varying}{Should the initial movement probabilities be time-varying? Default value is \code{TRUE}}

\item{spawn_yr}{spawn year to divide into weekly strata}

\item{spp}{choices are either \code{Chinook} or \code{Steelhead}}

\item{start_day}{date (\code{month / day}) when strata should start}

\item{end_day}{date (\code{month / day}) when strata should end}

\item{strata_beg}{3 letter code for the day of the week each weekly strata should begin on. Default value is \code{'Mon'}.}

\item{last_strata_min}{minimum length (in days) for the final strata. Default value is 3.}
}
\description{
Transforms data into appropriate format to feed into JAGS, including adding appropriate weekly strata for each tag if appropriate. Also constructs an initial value function to feed to JAGS. Currently this function is specific to the Lower Granite dam version of DABOM.
}
\examples{
\dontrun{prepJAGSinputs_LGD()}
}
\author{
Kevin See
}
