% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameTransParams_LGD.R
\name{renameTransParams_LGD}
\alias{renameTransParams_LGD}
\title{Rename DABOM Parameters - LGD}
\usage{
renameTransParams_LGD(param_vec = NULL)
}
\arguments{
\item{param_vec}{A character vector of transition parameter names}
}
\description{
Renames the various transition parameters in a matrix. This function is specific to the Lower Granite version of DABOM.
}
\examples{
renameTransParams_LGD()
}
\author{
Kevin See
}
