% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameTransParams_PRA.R
\name{renameTransParams_PRA}
\alias{renameTransParams_PRA}
\title{Rename DABOM Parameters - PRA}
\usage{
renameTransParams_PRA(param_vec = NULL)
}
\arguments{
\item{param_vec}{A character vector of transition parameter names}
}
\description{
Renames the various transition parameters in a matrix. This function is specific to the Priest Rapids version of DABOM.
}
\examples{
renameTransParams_PRA()
}
\author{
Kevin See
}
