% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameTransParams_TUM.R
\name{renameTransParams_TUM}
\alias{renameTransParams_TUM}
\title{Rename DABOM Parameters - TUM}
\usage{
renameTransParams_TUM(param_vec = NULL)
}
\arguments{
\item{param_vec}{A character vector of transition parameter names}
}
\description{
Renames the various transition parameters in a matrix. This function is specific to the Tumwater version of DABOM.
}
\examples{
renameTransParams_TUM()
}
\author{
Kevin See
}
