% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBranchNums.R
\name{setBranchNums}
\alias{setBranchNums}
\title{Set Branch Numbers}
\usage{
setBranchNums(parent_child = NULL)
}
\arguments{
\item{parent_child}{dataframe with at least \code{ParentNode}, \code{ChildNode} and \code{RKM} columns}
}
\description{
Returns list with the number of branches at each branching node based on parent child relationship table.
}
\examples{
setBranchNums()
}
\author{
Kevin See
}
