% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBranchNums_LGD.R
\name{setBranchNums_LGD}
\alias{setBranchNums_LGD}
\title{Set Branch Numbers - Lower Granite}
\usage{
setBranchNums_LGD()
}
\description{
Returns list with the number of branches at each branching node in the DABOM model for Lower Granite.
}
\examples{
setBranchNums_LGD()
}
\author{
Kevin See
}
