% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBranchNums_PRA.R
\name{setBranchNums_PRA}
\alias{setBranchNums_PRA}
\title{Set Branch Numbers - Priest}
\usage{
setBranchNums_PRA()
}
\description{
Returns list with the number of branches at each branching node in the DABOM model for Priest.
}
\examples{
setBranchNums_PRA()
}
\author{
Kevin See
}
