% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBranchNums_TUM.R
\name{setBranchNums_TUM}
\alias{setBranchNums_TUM}
\title{Set Branch Numbers - Tumwater}
\usage{
setBranchNums_TUM()
}
\description{
Returns list with the number of branches at each branching node in the DABOM model for Tumwater.
}
\examples{
setBranchNums_TUM()
}
\author{
Kevin See
}
