% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues.R
\name{setInitialValues}
\alias{setInitialValues}
\title{DABOM Initial Values}
\usage{
setInitialValues(filter_ch, parent_child, configuration)
}
\arguments{
\item{filter_ch}{filtered capture history as returned by the `filterDetections()`
function in the `PITcleanr` package, which has then been verified by a user and all
blank or NA `user_keep_obs` entries have been completed.}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`}
}
\value{
function
}
\description{
Construct appropriate initial values to be fed to JAGS. Focused on where tags were detected
}
\examples{
setInitialValues()
}
\author{
Kevin See
}
