% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues_PRA.R
\name{setInitialValues_PRA}
\alias{setInitialValues_PRA}
\title{LGD DABOM Initial Values}
\usage{
setInitialValues_PRA(dabom_list = NULL)
}
\arguments{
\item{dabom_list}{output of \code{createDABOMcapHist} with parameter split_matrics set to \code{TRUE}.}
}
\description{
Construct appropriate initial values for Lower Granite version of DABOM
}
\examples{
setInitialValues_PRA()
}
\author{
Kevin See
}
