% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues_PRO.R
\name{setInitialValues_PRO}
\alias{setInitialValues_PRO}
\title{PRO DABOM Initial Values}
\usage{
setInitialValues_PRO(dabom_list = NULL, model_file = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_list}{output of \code{createDABOMcapHist} with parameter split_matrics set to \code{TRUE}.}

\item{model_file}{file path to JAGS text file.}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Construct appropriate initial values for Prosser version of DABOM
}
\examples{
setInitialValues_PR)()
}
\author{
Kevin See
}
