% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues_TUM.R
\name{setInitialValues_TUM}
\alias{setInitialValues_TUM}
\title{Tumwater DABOM Initial Values}
\usage{
setInitialValues_TUM(dabom_list = NULL, n_branch_list = NULL)
}
\arguments{
\item{dabom_list}{output of \code{createDABOMcapHist} with parameter split_matrics set to \code{TRUE}.}

\item{n_branch_list}{list containing an entry for each branching node, and an integer indicating the number of branches upstream of that node.}
}
\description{
Construct appropriate initial values for Tumwater version of DABOM
}
\examples{
setInitialValues_TUM()
}
\author{
Kevin See
}
