% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSavedParams_LGD.R
\name{setSavedParams_LGD}
\alias{setSavedParams_LGD}
\title{Set Saved Parameters - LGD}
\usage{
setSavedParams_LGD(time_varying = F)
}
\arguments{
\item{time_varying}{Should the initial movement probabilities be time-varying? Default value is \code{FALSE}}
}
\description{
Returns a character vector of parameters to be saved by JAGS in the Lower Granite version of DABOM
}
\examples{
setSavedParams()
}
\author{
Kevin See
}
