% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSavedParams_PRA.R
\name{setSavedParams_PRA}
\alias{setSavedParams_PRA}
\title{Set Saved Parameters - PRA}
\usage{
setSavedParams_PRA(time_varying = F)
}
\arguments{
\item{time_varying}{Should the initial movement probabilities be time-varying? Default value is \code{FALSE}}
}
\description{
Returns a character vector of parameters to be saved by JAGS in the Priest Rapids version of DABOM
}
\examples{
setSavedParams_PRA()
}
\author{
Kevin See
}
