% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSavedParams_TUM.R
\name{setSavedParams_TUM}
\alias{setSavedParams_TUM}
\title{Set Saved Parameters - TUM}
\usage{
setSavedParams_TUM(time_varying = F)
}
\arguments{
\item{time_varying}{Should the initial movement probabilities be time-varying? Default value is \code{FALSE}}
}
\description{
Returns a character vector of parameters to be saved by JAGS in the Tumwater version of DABOM
}
\examples{
setSavedParams_TUM()
}
\author{
Kevin See
}
