% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawnerPaths.R
\name{spawnerPaths}
\alias{spawnerPaths}
\title{spawnerPaths: creates a data frame of all observed paths}
\usage{
spawnerPaths(valid_obs, valid_paths)
}
\arguments{
\item{valid_obs}{}

\item{valid_paths}{}

\item{tagCntLimit}{}
}
\description{
The function builds a data frame of all the observed fish paths in the
observation file based on the parent child table. It works by checking node observations
versus the previous node observation to see if the previous node was a parent or grandparent
great grand-parent.....
}
\examples{
spawnerPaths()

}
\author{
Ryan N. Kinzer
}

