% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validPaths.R
\name{validPaths}
\alias{validPaths}
\title{validPaths: creates a data frame of all available paths}
\usage{
validPaths(parent_child)
}
\arguments{
\item{parent_child}{is a data frame or comma seperated value (.csv) file containing the
parent-child nodes}
}
\description{
The function builds a data frame of all the available and acceptable paths
from the parent child table. These available paths are then used to check the observation
histories of each individual tagID.  The function was originally called
shortestPathToRoot_logCurrPaths() and was built by Greg K. Ryan K. altered the function
to work solely in R and doesn't require the SQLite backend.
}
\examples{
validPaths()

}
\author{
Greg Kliewer
}

