% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validTags.R
\name{validTags}
\alias{validTags}
\title{validTags: select valid tags from Lower Granite trapping data}
\usage{
validTags(trapdata)
}
\arguments{
\item{trapdata}{data frame outputted from LGTrapData() function}
}
\description{
validTags() is a function to subset only valid tags
from the LGTrappingExportJodyW.accdb database. A dataframe,
outputted from the LGTrapData() function, is supplied and then trimmed
for only those records deemed valid (by IDFG), being AD-Intact and having
a PIT-tag number.  Chinook are also trimmed to only those fish with a run
designation of '5' in the SRR field.
}
\examples{
validTags(trapdata = trap_df)

}
\author{
Ryan N. Kinzer
}

