% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDABOM.R
\name{writeDABOM}
\alias{writeDABOM}
\title{Write DABOM JAGS model}
\usage{
writeDABOM(file_name = NULL, parent_child = NULL, configuration = NULL)
}
\arguments{
\item{file_name}{name (with file path) to save the model as}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`}
}
\description{
This writes the overall JAGS model for a generic DABOM as a text file. It can then be modified depending on the observations for a particular valid tag list.
}
\examples{
writeDABOM()
}
\author{
Kevin See
}
