% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDABOM_LGD.R
\name{writeDABOM_LGD}
\alias{writeDABOM_LGD}
\title{Write Lower Granite Dam JAGS model}
\usage{
writeDABOM_LGD(file_name = NULL)
}
\arguments{
\item{file_name}{name (with file path) to save the model as}
}
\description{
This writes the overall JAGS model for LGR DABOM as a text file. It can then be modified depending on the observations for a particular valid tag list.
}
\examples{
writeJAGSmodel()
}
\author{
Kevin See
}
