% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDABOM_PRA.R
\name{writeDABOM_PRA}
\alias{writeDABOM_PRA}
\title{Write Priest Rapids JAGS model}
\usage{
writeDABOM_PRA(file_name = NULL)
}
\arguments{
\item{file_name}{name (with file path) to save the model as}
}
\description{
This writes the overall JAGS model for Priest Rapids DABOM as a text file. It can then be modified depending on the observations for a particular valid tag list.
}
\examples{
writeJAGSmodel_PRA()
}
\author{
Kevin See
}
