% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDABOM_PRO.R
\name{writeDABOM_PRO}
\alias{writeDABOM_PRO}
\title{Write Prosser JAGS model}
\usage{
writeDABOM_PRO(file_name = NULL)
}
\arguments{
\item{file_name}{name (with file path) to save the model as}
}
\description{
This writes the overall JAGS model for Prosser Dam DABOM as a text file. It can then be modified depending on the observations for a particular valid tag list.
}
\examples{
writeJAGSmodel_PRO()
}
\author{
Kevin See
}
