% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDABOM_TUM.R
\name{writeDABOM_TUM}
\alias{writeDABOM_TUM}
\title{Write Tumwater JAGS model}
\usage{
writeDABOM_TUM(file_name = NULL)
}
\arguments{
\item{file_name}{name (with file path) to save the model as}
}
\description{
This writes the overall JAGS model for Tumwater DABOM as a text file. It can then be modified depending on the observations for a particular valid tag list.
}
\examples{
writeJAGSmodel()
}
\author{
Kevin See
}
