% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAbundPost.R
\name{calcAbundPost}
\alias{calcAbundPost}
\title{Calculate Abundance Posteriors}
\usage{
calcAbundPost(
  move_post = NULL,
  abund_post = NULL,
  bootstrap_samp = 2000,
  .move_vars = c("origin", "param"),
  .abund_vars = c("origin"),
  time_vary_param_nm = NULL
)
}
\arguments{
\item{bootstrap_samp}{The number of samples to be drawn from the posteriors of the STADEM model and the DABOM model.}

\item{time_vary_param_nm}{column name describing the time-varying strata for each location. Default value is `strata_num`.}
}
\description{
Combines estimates of total escapement from STADEM with transition probabilities from DABOM to generate estimates of escapement above various detection sites.
}
\examples{
#calcAbundPost()
}
\author{
Kevin See
}
