% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRepGrpEscape_LGD.R
\name{calcRepGrpEscape_LGD}
\alias{calcRepGrpEscape_LGD}
\title{Estimate Escapement for Reporting Groups - LGD}
\usage{
calcRepGrpEscape_LGD(
  dabom_mod = NULL,
  stadem_mod = NULL,
  stadem_param_nm = "X.new.wild",
  bootstrap_samp = 2000,
  node_order = NULL,
  summ_results = T,
  pt_est_nm = NULL,
  cred_int_prob = 0.95,
  spp = c("Chinook", "Steelhead")
)
}
\arguments{
\item{spp}{Species, either "Chinook" or "Steelhead"}
}
\description{
Combines estimates of total escapement from STADEM with transition probabilities from DABOM to generate estimates of escapement within pre-defined reporting groups. Currently only works for wild fish.
}
\examples{
#calcRepGrpEscape_LGD()
}
\author{
Kevin See
}
