% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTransProbs_PRA.R
\name{compileTransProbs_PRA}
\alias{compileTransProbs_PRA}
\title{Compile Transition Probabilities - PRA}
\usage{
compileTransProbs_PRA(dabom_mod = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model, and multiplies them appropriately. This function is specific to the Lower Granite version of DABOM.
}
\examples{
compileTransProbs_PRA()
}
\author{
Kevin See
}
