% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTransProbs_TUM.R
\name{compileTransProbs_TUM}
\alias{compileTransProbs_TUM}
\title{Compile Transition Probabilities - TUM}
\usage{
compileTransProbs_TUM(dabom_mod = NULL)
}
\arguments{
\item{dabom_mod}{An MCMC.list}
}
\description{
Extracts the MCMC posteriors of transition probabilities for a DABOM model, and multiplies them appropriately. This function is specific to the Lower Granite version of DABOM.
}
\examples{
compileTransProbs_TUM()
}
\author{
Kevin See
}
