% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJAGSinputs_TUM.R
\name{createJAGSinputs_TUM}
\alias{createJAGSinputs_TUM}
\title{Prep DABOM JAGS inputs - TUM}
\usage{
createJAGSinputs_TUM(dabom_list = NULL, model_file = NULL, parent_child = NULL)
}
\arguments{
\item{dabom_list}{output of \code{createDABOMcapHist} with parameter split_matrics set to \code{TRUE}.}

\item{model_file}{file path to JAGS text file.}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Construct all the necessary inputs to the DABOM JAGS model for TUM. Returns a named list that can be passed directly to JAGS.
}
\examples{
createJAGSinputs_TUM()
}
\author{
Kevin See
}
