% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineDabomColNms.R
\name{defineDabomColNms}
\alias{defineDabomColNms}
\title{Define Column Names for DABOM matrix}
\usage{
defineDabomColNms(
  root_site = NA,
  parent_child,
  configuration,
  second_node = TRUE
)
}
\arguments{
\item{root_site}{determines which version of DABOM the user is running.}

\item{parent_child}{parent-child table. Could be created from `buildParentChild()` from `PITcleanr` package.}

\item{configuration}{configuration file. Could be created from `buildConfig()` from `PITcleanr` package.}

\item{second_node}{Default option is `TRUE` which defines the `bottom_sites` as the second order nodes above the `root_site`, (i.e., `node_order == 2`). If `FALSE`, some sites are hard-coded depending on what the `root_site` is set to.  Default is `TRUE`; otherwise `bottom_sites` should be defined within the `defineDabomColNms()`
function}
}
\description{
based on a parent-child table, this returns a vector of nodes to help create a consistent
series of DABOM input matrices.
}
\examples{
defineDabomColNms()
}
\author{
Kevin See
}
