% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineRepGrps.R
\name{defineRepGrps}
\alias{defineRepGrps}
\title{Define Report Groups - LGD}
\usage{
defineRepGrps(spp = c("Chinook", "Steelhead"), node_order = NULL)
}
\arguments{
\item{spp}{Species, either "Chinook" or "Steelhead"}

\item{node_order}{output of function \code{createNodeOrder}.}
}
\description{
Define which detection sites fall into which report groups, and what the lowest array is. These reporting gropus differ by species.
}
\examples{
defineRepGrps()
}
\author{
Kevin See
}
