% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNoFishNodes.R
\name{fixNoFishNodes}
\alias{fixNoFishNodes}
\title{Modify basic JAGS model}
\usage{
fixNoFishNodes(
  init_file = NULL,
  file_name = NULL,
  filter_ch = NULL,
  parent_child = NULL,
  configuration = NULL
)
}
\arguments{
\item{init_file}{name (with file path) of basic JAGS model.}

\item{file_name}{name (with file path) to save the model as.}

\item{filter_ch}{filtered capture history as returned by the `filterDetections()`
function in the `PITcleanr` package, which has then been verified by a user and all
blank or NA `user_keep_obs` entries have been completed.}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`}
}
\description{
Takes the basic JAGS model and modifies it based on particular observations. Specifically, it fixes the detection probability of all nodes that had no observations to 0.
}
\examples{
fixNoFishNodes()
}
\author{
Kevin See
}
