% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBranchNums.R
\name{setBranchNums}
\alias{setBranchNums}
\title{Set Branch Numbers}
\usage{
setBranchNums(parent_child = NULL)
}
\arguments{
\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function in the `PITcleanr` package.}
}
\description{
Returns list with the number of branches at each branching node based on parent child relationship table.
}
\examples{
setBranchNums()
}
\author{
Kevin See
}
