% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePost.R
\name{summarisePost}
\alias{summarisePost}
\title{Summarize Posteriorrs}
\usage{
summarisePost(.data, value, ..., .cred_int_prob = 0.95)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble) containing a column called `value`}

\item{value}{name of the column containing posterior values. Default is `value`.}

\item{...}{<tidy-select> Columns to group by (e.g. `origin`, `param`, etc.)
#' @param .cred_int_prob A numeric scalar in the interval (0,1) giving what higest posterior density portion of the posterior the credible interval should cover. The default value is 95\%.}
}
\description{
After the user has extracted posterior draws from an `mcmc` object and wrangled them into a long tibble, this function computes several summary statistics.
}
\examples{
#summarisePost()
}
\author{
Kevin See
}
