% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validTagList.R
\name{validTagList}
\alias{validTagList}
\title{validTagList: Generate valid PIT tag list from LGTrappingDB}
\usage{
validTagList(input = NULL, spawnYear = NULL, species = "chnk",
  exportFile = NULL)
}
\arguments{
\item{input}{the name of the tab-delimited text file containing all of the records from the LGTrappingDB.}

\item{spawnYear}{the desired spawn year}

\item{species}{which species?}

\item{exportFile}{name of export file?}
}
\description{
validTagList() is a function to generate a valid tag list using the LGTrappingDB.
}
\examples{
validTagList(input = LGTrappingDB_03192017, species )

}
\author{
Mike Ackerman
}

