% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validTagList.R
\name{validTagList}
\alias{validTagList}
\title{validTagList: Generate valid PIT tag list from LGTrappingDB}
\usage{
validTagList(input = NULL, spawnYear = "SY2016", species = "chnk",
  validTagFile = "validTagFile.txt", validTagData = "validTagData.csv")
}
\arguments{
\item{input}{the name of the tab-delimited text file containing all of the records from the LGTrappingDB.}

\item{spawnYear}{Which spawn year do you want to pull records for? Should be in character format. For example, to
pull records for spawn year 2016 use \code{"SY2016"}.}

\item{species}{which species?}

\item{validTagFile}{name of valid tag list file? Will be exported as tab delimited .txt}

\item{validTagData}{name of file for all data associated with the valid tag list}
}
\description{
validTagList() is a function to generate a valid tag list using the LGTrappingDB.
}
\examples{
validTagList(input = LGTrappingDB_03192017, spawnYear = "SY2016", species = "chnk", exportFile = "sy2016chnkValidTagList.txt") )

}
\author{
Mike Ackerman
}

